Exposing the result of input value preparation as a FieldValue instance
This change affects the result type of Evaluator#prepare(FieldName,
Object) method. Despite the fact that the result type is not part of
method signature, it breaks the binary compatibility with earlier
versions. The source compatibility is preserved, though, because the new
result type FieldValue is a subclass or earlier result type Object.