Inference of the SVM classification method type
The SupportVectorMachineModel element specifies an optional attribute
'classificationMethod', which informs the PMML consumer if the
classification should be performed using one-against-all (the default)
or one-against-one approach.

Currently, several PMML producers disregard this attribute (ie. leave it
uninitialized in situations where it should be initialized with a
non-default value). The JPMML library is able to automatically identify
and correct such deficiencies during model evaluation.